refactor: MyReservationResponse를 reservation 패키지로 이동
