+++ src/main/java/roomescape/reservation/ReservationService.java
@@ -4,7 +4,6 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import roomescape.member.LoginMember;
 import roomescape.member.Member;
-import roomescape.member.MyReservationResponse;
 import roomescape.theme.Theme;
 import roomescape.theme.ThemeRepository;
 import roomescape.time.Time;
